버튼 기능 구현(db 연결용 코드 포함) 완료, 수정창 드롭다운메뉴 수정 중,디데이 계산 중

+++ InTravel/app/src/main/java/com/example/intravel/data/MainData.kt
@@ -1,6 +1,6 @@
 package com.example.intravel.data
 
-data class MainData(var tId:Int,
+data class MainData(var tId:Long,
                     var tTitle: String,
                     var createDate:String,
                     var startDate:String,
